requirements.txt
asgiref==3.10.0
openai==2.8.1
beautifulsoup4==4.14.2
djangorestframework==3.16.1
Django==5.2
django-htmlmin==0.11.0
django-ipware==7.0.1
django-jazzmin==3.0.1
html5lib==1.1
psycopg2-binary==2.9.11
python-decouple==3.8
python-ipware==3.0.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
webencodings==0.5.1
